Moving Path related functions to Paths.py
